+++ packages/database/database.ts
@@ -2,7 +2,11 @@
 import { Tracer, Span } from "@fwl/tracing";
 import { Pool, QueryArrayResult, PoolClient } from "pg";
 
-import { DatabaseConfig, defaultConfig } from "./config/config";
+import {
+  DatabaseConfig,
+  DatabaseConfigWithDatabaseUrl,
+  defaultConfig,
+} from "./config/config";
 
 function close(pool: Pool): Promise<void> {
   return pool.end();
@@ -235,22 +239,26 @@ function checkDatabaseError(error: any): void {
 
 export function connect(
   tracer: Tracer,
-  options?: DatabaseConfig,
+  options?: DatabaseConfig | DatabaseConfigWithDatabaseUrl,
 ): DatabaseQueryRunner {
-  const config = options ? options : defaultConfig;
+  const config = getConfig(options);
+
   const pool = new Pool({
     user: config.username,
     password: config.password,
     host: config.host,
     database: config.database,
     port: config.port,
+    ssl: config.ssl,
   });
   return queryRunner(pool, tracer);
 }
 
 export function connectWithoutTracing(
-  config: DatabaseConfig,
+  options?: DatabaseConfig | DatabaseConfigWithDatabaseUrl,
 ): DatabaseQueryRunnerWithoutTracing {
+  const config = getConfig(options);
+
   const pool = new Pool({
     user: config.username,
     password: config.password,
@@ -262,8 +270,10 @@ export function connectWithoutTracing(
 }
 
 export async function connectInSandbox(
-  config: DatabaseConfig,
+  options?: DatabaseConfig | DatabaseConfigWithDatabaseUrl,
 ): Promise<DatabaseQueryRunnerSandbox> {
+  const config = getConfig(options);
+
   const pool = new Pool({
     user: config.username,
     password: config.password,
@@ -274,4 +284,25 @@ export async function connectInSandbox(
   const client: PoolClient = await pool.connect();
   await client.query("BEGIN");
   return queryRunnerSandbox(pool, client);
+}
+
+function convertUrl(databaseUrl: string): DatabaseConfig {
+  const { hostname, port, password, username, pathname } = new URL(databaseUrl);
+  return {
+    host: hostname,
+    username,
+    password,
+    port: parseInt(port),
+    database: pathname.slice(1),
+  };
+}
+
+function getConfig(
+  options?: DatabaseConfig | DatabaseConfigWithDatabaseUrl,
+): DatabaseConfig {
+  return options
+    ? "url" in options
+      ? { ...convertUrl(options.url), ssl: options.ssl || false }
+      : { ...defaultConfig, ...options }
+    : defaultConfig;
 }