+++ packages/web/src/next/index.ts
@@ -1,4 +1,9 @@
-import { GetServerSidePropsContext, GetServerSidePropsResult } from "next";
+import {
+  GetServerSidePropsContext,
+  GetServerSidePropsResult,
+  NextApiRequest,
+  NextApiResponse,
+} from "next";
 
 import { wrapMiddlewares } from "../middlewares/wrapper";
 import { Handler } from "../typings/handler";
@@ -6,7 +11,7 @@ import { Middleware } from "../typings/middleware";
 
 export async function getServerSidePropsWithMiddlewares<P>(
   context: GetServerSidePropsContext,
-  middlewares: Middleware[],
+  middlewares: Middleware<NextApiRequest, NextApiResponse>[],
   handler: Handler = () => Promise.resolve(),
 ): Promise<GetServerSidePropsResult<P>> {
   const result = await wrapMiddlewares(middlewares, handler)(